Introduce UpdateMasterPasswordAction
