+++ src/main.rs
@@ -2,7 +2,6 @@ extern crate clipboard;
 #[macro_use]
 extern crate magic_crypt;
 use crate::auth::AccessTokens;
-use anyhow::Context;
 use clap::{arg, ArgAction, Command};
 
 use crate::actions::Action;
@@ -16,7 +15,6 @@ mod online_vault;
 mod password;
 mod store;
 mod ui;
-use log::debug;
 use std::env;
 use std::io;
 
@@ -96,18 +94,7 @@ async fn main() -> anyhow::Result<()> {
         Some(("show", sub_matches)) => actions::ShowAction::new(sub_matches).execute().await?,
         Some(("delete", sub_matches)) => actions::DeleteAction::new(sub_matches).execute().await?,
         Some(("csv", sub_matches)) => actions::ImportCsvAction::new(sub_matches).execute().await?,
-        Some(("password", _)) => {
-            let old_pwd = ui::ask_master_password("Enter current master password: ".into());
-            let new_pwd = ui::ask_new_password();
-            let success = update_master_password(&old_pwd, &new_pwd)
-                .await
-                .context("Failed to update master password")?;
-            if success {
-                println!("Password changed");
-            } else {
-                println!("Failed to change master password");
-            }
-        }
+        Some(("password", _)) => actions::UpdateMasterPasswordAction {}.execute().await?,
         Some(("keychain-push", _)) => {
             let master_pwd = ui::ask_master_password(None);
             let creds = store::get_all_credentials();
@@ -135,18 +122,4 @@ async fn main() -> anyhow::Result<()> {
         }
     }
     Ok(())
-}
-
-async fn update_master_password(old_pwd: &str, new_pwd: &str) -> anyhow::Result<bool> {
-    if store::has_logged_in() {
-        debug!("Updating master password in online vault!");
-        let token = actions::get_access_token().await?;
-        let count =
-            online_vault::update_master_password(&token.access_token, old_pwd, new_pwd).await?;
-        store::save_master_password(new_pwd);
-        debug!("Updated {} passwords", count);
-    } else {
-        store::update_master_password(old_pwd, new_pwd);
-    }
-    Ok(true)
 }
